requirements.txt
kagglehub
python-dotenv
openai
anthropic
google-genai
together
mistralai
